Initialize Back-end

+++ database/src/main/java/com/cookguide/database/service/AccountService.java
@@ -1,6 +1,7 @@
 package com.cookguide.database.service;
 
 import com.cookguide.database.model.Account;
+import com.cookguide.database.repository.AccountRepository;
 
 import java.util.List;
 
@@ -8,8 +9,8 @@ public interface AccountService {
     public abstract Account createAccount(Account account);
 
     public abstract Account updateAccount(Account account);
-    public abstract Account getAccountById(int id);
+    public abstract Account getAccountById(Long id);
     public abstract List<Account> getAllAccounts();
-    public abstract void deleteAccount(int id);
-    public abstract boolean isAccountExist(int id);
+    public abstract void deleteAccount(Long id);
+    public abstract boolean isAccountExist(Long id);
 }
